backend/pyproject.toml
[project]
name = "sentineliam"
version = "0.1.2"
description = "SentinelIAM CLI for policy generation, validation, and audit"
authors = [{ name = "Katrina Finney", email = "you@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "typer[all]",
  "jinja2",
  "pydantic",
  "pyyaml",
  "supabase",
  "python-dotenv"
]

[project.scripts]
sentinel = "cli.main:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

